enhanced_requirements.txt
# Enhanced Dinosaur DNA Reconstruction System
# Optimized for Linux/WSL environment

# Core ML and Deep Learning
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.24.0

# DNA/Biology specific models
dnabert-pytorch>=0.1.0
bio-transformers>=0.1.0
biopython>=1.81
scikit-bio>=0.5.8

# Sequence Analysis and Phylogenetics
dendropy>=4.6.1
ete3>=3.1.3
pyvolve>=0.8.5
cogent3>=2023.5.17a6
phylopandas>=0.8.0

# Scientific Computing (Linux optimized)
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.1.0
scikit-learn>=1.3.0

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
bokeh>=3.3.0

# Progress and Monitoring
tqdm>=4.66.0
wandb>=0.16.0
tensorboard>=2.15.0

# Web Interfaces
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
gradio>=4.7.0

# Data Processing
requests>=2.31.0
aiohttp>=3.9.0
pyyaml>=6.0.1
jsonlines>=4.0.0

# System utilities (Linux/WSL specific)
psutil>=5.9.0
py-cpuinfo>=9.0.0
gputil>=1.4.0

# Development and Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0

# Memory and Performance
memory-profiler>=0.61.0
line-profiler>=4.1.0
numba>=0.58.0

# Error handling and logging
rich>=13.7.0
loguru>=0.7.2

# File handling
h5py>=3.10.0
zarr>=2.16.0
